major commit : integrate BeatNet into project. Now BeatNet is set as the default tracker for dev purposes.
Technical dept: BeatNet computes upbeats and downbeats. To calculate the tempo, the calculate tempo function has to be implemented. This will compute the tempo based on the beat interval between two adjusted beats.

Techincal dept 2: After building the plugin, all DLL files and model weights must be manually copied to the application host's installation directory. This is a temporary manual fix due to the current state of development.